requirements.txt
asgiref==3.8.1
Django==5.0.6
django-mssql-backend==2.8.1
djangorestframework==3.15.2
mysql-connector-python==8.4.0
mysqlclient==2.2.4
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pyodbc==5.1.0
pytesseract==0.3.10
sqlparse==0.5.0
tesseract==0.1.3
tzdata==2024.1
